reorganised js files into client & server folders
