+++ www/index.html
@@ -21,10 +21,10 @@
 
 		<!-- Locally Sourced
 			<link rel="stylesheet" href="./css/bootstrap.min.css">
-			<script src="./js/jquery-3.5.1.min.js"></script>
-			<script src="./js/popper.min.js"></script>
-			<script src="./js/bootstrap.min.js"></script>
-			<script src="./js/cytoscape.min.js"></script>
+			<script src="./clientSide/jquery-3.5.1.min.js"></script>
+			<script src="./clientSide/popper.min.js"></script>
+			<script src="./clientSide/bootstrap.min.js"></script>
+			<script src="./clientSide/cytoscape.min.js"></script>
 
 			
 		-->
@@ -32,14 +32,14 @@
 		<link rel="stylesheet" href="./css/style.css">
 
 		<!-- Page scripts -->
-		<script src="./classes/Node.js" type="text/javascript" defer></script>
-		<script src="./classes/NodeModal.js" type="text/javascript" defer></script>
-		<script src="./classes/GraphSettings.js" type="text/javascript" defer></script>
+		<script src="./clientSide/Node.js" type="text/javascript" defer></script>
+		<script src="./clientSide/NodeModal.js" type="text/javascript" defer></script>
+		<script src="./clientSide/GraphSettings.js" type="text/javascript" defer></script>
 
-		<script src="./js/helpers.js" type="text/javascript" defer></script>
-		<script src="./js/modals.js" type="text/javascript" defer></script>
-		<script src="./js/index.js" type="text/javascript" defer></script>
-		<script src="./js/events.js" type="text/javascript" defer></script>
+		<script src="./clientSide/helpers.js" type="text/javascript" defer></script>
+		<script src="./clientSide/modals.js" type="text/javascript" defer></script>
+		<script src="./clientSide/index.js" type="text/javascript" defer></script>
+		<script src="./clientSide/events.js" type="text/javascript" defer></script>
 
 		<title>System of Systems Modeller</title>
